тесты для селектора offer
